refactor: extract workspace transfer service
